+++ components/BtnWithBusinessFormModal/BtnWithBusinessFormModal.tsx
@@ -6,7 +6,7 @@ import { ModalWindow } from '@/components/ui/ModalWindow';
 import { BusinessForm } from '@/components/BusinessForm/BusinessForm';
 import { SuccessNotification } from '@/components/SuccessNotification/SuccessNotification';
 import { ErrorNotification } from '@/components/ErrorNotification/ErrorNotification';
-import { BtnWithBusinessFormModalProps, PopUpType } from './types';
+import { BtnWithBusinessFormModalProps, PopUpType } from '@/types';
 
 export const BtnWithBusinessFormModal: FC<BtnWithBusinessFormModalProps> = ({
   staticData,
