--- components/ButtonLoader/types.ts
@@ -1,4 +1,0 @@
-export type ButtonLoaderProps = {
-  color: string;
-  className: string;
-};
