+++ types/index.ts
@@ -1,1 +1,4 @@
 export * from '../layout/Footer/types';
+
+//business form
+export * from './BusinessFormTypes';
